Background loop + lander + human
Set up the core background that loops to the other side and set up cinemachine for following player and confining map. Set up basic idle animations and state machines for a human and lander enemy. Fixed bug where projectile didnt flip correctly when the player ship was facing the left.